# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "createProductSchema": {
        "type": "object",
        "required": [
          "name",
          "stock",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Name of the product",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Detailed description of the product",
            "example": "High-quality bluetooth headphones with noise cancellation."
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "Category this product belongs to",
            "example": "Electronics"
          },
          "stock": {
            "type": "number",
            "description": "Available product stock",
            "example": 50
          },
          "price": {
            "type": "number",
            "description": "Price of the product",
            "example": 129.99
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "description": "Whether the product is active and visible",
            "example": true
          }
        }
      },
      "updateProductSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Updated product name",
            "example": "Updated Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Updated description of the product",
            "example": "New improved model with longer battery life."
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "Updated category of the product",
            "example": "Audio"
          },
          "stock": {
            "type": "number",
            "nullable": true,
            "description": "Updated stock number (if provided)",
            "example": 80
          },
          "price": {
            "type": "number",
            "nullable": true,
            "description": "Updated price (if provided)",
            "example": 149.99
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active",
            "example": true
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admin/setCustomClaims": {
      "post": {
        "summary": "Set custom claims for a specific user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "This endpoint allows an admin to assign custom claims (roles or permissions)   to a Firebase Authentication user. Only users with an \"admin\" role can access this route.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "uid",
                  "claims"
                ],
                "properties": {
                  "uid": {
                    "type": "string",
                    "example": "aBcD1234EfGhIj",
                    "description": "The Firebase user UID to update."
                  },
                  "claims": {
                    "type": "object",
                    "example": {
                      "role": "manager"
                    },
                    "description": "Custom claims object to assign to the user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom claims were successfully set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Custom claims set for user: aBcD1234EfGhIj"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Unauthorized — missing or invalid token."
          },
          "403": {
            "description": "Forbidden — user does not have admin privileges."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/products": {
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createProductSchema"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "images": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "binary"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all products",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Fetched all products successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a product by ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/updateProductSchema"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "images": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "binary"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a product by ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}